Revert "Revert "宝宝心里苦啊……重新写了函数""
This reverts commit 40a0d286cf370490e665f8a4e12eeb7ed9176781.

+++ Strategy/Strategy.h
@@ -24,7 +24,7 @@ int TreePolicy(int i);
 
 int Expand(int i);
 
-int BestChild(int v, int c);
+int BestChild(int v, double c);
 
 double DefaultPolicy(int v);
